requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
playwright
atlassian-python-api
